project_django/requirements.txt
Django~=4.1.1
django-allauth~=0.51.0
django-filter~=22.1
celery~=5.2.7
redis~=4.3.4
django_redis~=5.2.0